+++ packages/client-core/src/networking/NetworkInstanceProvisioning.tsx
@@ -40,7 +40,6 @@ export const NetworkInstanceProvisioning = () => {
   const currentChannelInstanceConnection = useMediaInstance()
 
   MediaInstanceConnectionService.useAPIListeners()
-  PartyService.useAPIListeners()
 
   useRoomCodeURLParam(false, true)
